Closes #64
