+++ api/odk/osm/osm2osc.js
@@ -1,0 +1,5 @@
+
+module.exports = function(osmXmlStr) {
+    
+    return "";
+};
